Renamed mcols to column_data.
Currently, the length of a BiocFrame is the number of rows, while the mcols
contain metadata on the columns. This means that we violate the basic contract
of mcols, i.e., that it is parallel to the object length. (Indeed, we were
already verging on violating it anyway because the BiocFrame mcols needs to be
None-able to avoid an infinite recursion during construction.)

So, the simple solution is to just rename it to column_data, which frees us
from the mcols contract and is arguably more descriptive for a 2-dimensional
object. This change aligns us with other 2D classes like SummarizedExperiment.